NttIntt16Test(debug), trlgsw(wip), add second lvl decompose support for trgsw(wip), review todos, test decompose/ntt order (should have no diff), BlindRotLut(drift improve / wip)

+++ yatfhe/include/yatfhe/torus.h
@@ -13,7 +13,7 @@
 
 #ifdef TORUS32
 using Torus = int32_t; // use 32-bit int as torus to handle modular arithmetic naturally
-using Ntt16 = uint16_t;
+using Ntt16 = uint32_t;
 using Ntt32 = uint32_t;
 using Ntt64 = uint64_t;
 using UnsignedInteger = uint32_t;
